Add files via upload
The purpose of this exercise is to demonstrate how the cross-correlation function can be used
for finding the blood velocity in the human body.

1. Write a Matlab script to calculate the cross-correlation function between two signals, and
plot the cross-correlation function with the correct axis. Use the function xcorr and
employ the correct scaling.
Make a random signal by convolving a 2 cycle sinusoid with a white, random signal
generated by the procedure randn. The sampling frequency should be 100 MHz, the
frequency of the sinusoid 5 MHz, and the length of the random signal should be 20
microseconds. Make a delayed version of the signal by adding zeros at the start of the
signals, so that the delay corresponds to 0.4 microseconds.
Correlate the two random signals and find the delay from the maximum value in the
cross-correlation function.
Write a function that automatically determines the delay from the two signals and the
relevant parameters (Hint: see help max). The delay must be given in seconds.
2. Finding the delay between two random signals can be used for finding the blood velocity
in medical ultrasound. Here two signals are measured and cross-correlated to find the
delay between the signals. The time between the measurements is Tprf (pulse repetition
frequency period) and the delay ts found from the cross-correlation function is related to
the blood velocity v by:
ts
= 2 v Tprf / c
where c is the speed of sound (1540 m/s in human tissue). Determining the delay, thus,
directly reveals the velocity.
Make a function that determines the blood velocity from two measurements. The input
parameters to your function should be 1) the two signals, 2) Tprf, 3) c, and 4) the sampling
frequency fs.
The output should be the blood velocity in meters per second.
3. Load the file EX1_plug_flow.mat into Matlab with load <file_name>. The Matlab
variable data holds ten pulse-echo signals measured with fs=100 MHz, Tprf =200
microseconds, and c=1540 m/s.
Modify your routine so it can find the averaged cross-correlation function over the 10
lines. This is done by cross-correlating line 1 with 2, line 2 with 3 and so forth, and then
average all the cross-correlation functions. The maximum is then found from the
combined cross-correlation function.
Use your routine to find the velocity of the blood.
4. Divide the RF data into small segments of 2 microseconds and find the velocities within
the segments. Incorporate the division into your function, so it returns an array with the
estimated values.
5. Load the file EX1_carotis.mat into Matlab with load <file_name>. The Matlab variable
data holds 20 pulse-echo signals measured with fs=100 MHz, Tprf =200 microseconds,
and c=1540 m/s. Use your routine to find the velocity of the blood for 2 microsecond
segments. Plot the velocity as a function of segment number or depth, and observe how
the velocity changes as a function of depth (segment number). Remember that the depth d
is given by
d = t c/2,
where t is the time for the signal